Completed assignment

+++ Binomial.java
@@ -18,9 +18,9 @@ public class Binomial {
     }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		if (k > n) return 0;
-        int[][] memo = new int[n + 1][k + 1];
+        long[][] memo = new long[n + 1][k + 1];
         for (int i = 0; i < memo.length; i++) {
             for (int j = 0; j < memo[i].length; j++) {
                 memo[i][j] = -1;
@@ -29,16 +29,14 @@ public class Binomial {
         return binomial(n, k, memo);
     }
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
-		// Base case
 		if ((k > n)) {
 		   	memo[n][k] = 0; 
 		   	return 0;
 		}
-		// Another base case
 		if (n == 0 || k == 0) {
 		   	memo[n][k] = 1; 
 		   	return 1;
